Move view tools to smaller components
