+++ app/src/components/feature/leftpanel/studyinfo/StudyInfo.component.jsx
