+++ app/src/components/feature/leftpanel/viewtools/ViewTools.component.jsx
@@ -1,0 +1,161 @@
+import { Button } from '@chakra-ui/button';
+import { Heading, HStack, VStack } from '@chakra-ui/layout';
+import { IconButton, Tooltip, useColorMode } from '@chakra-ui/react';
+import { Anchor, Bolt, Undo } from 'css.gg';
+import { observer } from 'mobx-react';
+import { useContext, useEffect, useState } from 'react';
+import CanvasTools from 'components/feature/leftpanel/viewtools/canvastools/CanvasTools.component';
+import EdgeTools from 'components/feature/leftpanel/viewtools/edgetools/EdgeTools.component';
+import NodeTools from 'components/feature/leftpanel/viewtools/nodetools/NodeTools.component';
+import FilterTools from 'components/feature/leftpanel/viewtools/filtertools/FilterTools.component';
+
+import { RootStoreContext } from 'stores/RootStore';
+
+function ViewTools() {
+    const store = useContext(RootStoreContext);
+    const [forceRunning, setForceRunning] = useState(false);
+    const { colorMode } = useColorMode();
+
+    useEffect(() => {
+        if (!store.graphInstance.forceEngine) {
+            setForceRunning(false);
+        }
+    }, [store.graphInstance.forceEngine]);
+
+    const renderLayoutOptions = () => {
+        return (
+            <VStack width="100%">
+                <HStack width="100%">
+                    <Tooltip label="Simulate graph layout (will make nodes move around and make it easier to see patterns)">
+                        <Button
+                            id="applyforcebutton"
+                            size="sm"
+                            leftIcon={<Bolt style={{ '--ggs': '0.6' }} />}
+                            backgroundColor={
+                                forceRunning
+                                    ? 'blue.400'
+                                    : colorMode === 'light'
+                                    ? 'blackAlpha.200'
+                                    : 'whiteAlpha.200'
+                            }
+                            onClick={() => {
+                                if (forceRunning) {
+                                    store.graphInstance.stopForce();
+                                    setForceRunning(false);
+
+                                    store.track.trackEvent(
+                                        'Side panel - View Settings',
+                                        'Button',
+                                        JSON.stringify({
+                                            type: 'Click',
+                                            value: 'Run Force'
+                                        })
+                                    );
+                                } else {
+                                    store.graphInstance.applyForce();
+                                    setForceRunning(true);
+                                    store.track.trackEvent(
+                                        'Side panel - View Settings',
+                                        'Button',
+                                        JSON.stringify({
+                                            type: 'Click',
+                                            value: 'Stop Force'
+                                        })
+                                    );
+                                }
+                            }}
+                            width="100%"
+                        >
+                            {forceRunning ? 'Stop force' : 'Run Force'}
+                        </Button>
+                    </Tooltip>
+                    <Tooltip
+                        label={
+                            store.graphInstance.forceShouldIgnoreSelected
+                                ? 'Turn off layout position simulation for currently selected nodes.'
+                                : 'Turn on layout position simulation for currently selected nodes.'
+                        }
+                    >
+                        <IconButton
+                            id="resetLayoutButton"
+                            size="sm"
+                            icon={<Anchor style={{ '--ggs': '0.6' }} />}
+                            onClick={() => {
+                                store.graphInstance.ignoreSelected(
+                                    !store.graphInstance
+                                        .forceShouldIgnoreSelected
+                                );
+                            }}
+                            _hover={{
+                                backgroundColor: 'blue.500',
+                                opacity: 0.76
+                            }}
+                            backgroundColor={
+                                store.graphInstance.forceShouldIgnoreSelected
+                                    ? 'blue.500'
+                                    : 'auto'
+                            }
+                            disabled={store.graphInstance.forceEngine}
+                        />
+                    </Tooltip>
+                    <Tooltip label="Reset node positions">
+                        <IconButton
+                            id="resetLayoutButton"
+                            size="sm"
+                            icon={<Undo style={{ '--ggs': '0.6' }} />}
+                            onClick={() => {
+                                store.graph.resetNodesPositions();
+
+                                store.track.trackEvent(
+                                    'Side panel - View Settings',
+                                    'Button',
+                                    JSON.stringify({
+                                        type: 'Click',
+                                        value: 'Reset layout'
+                                    })
+                                );
+                            }}
+                            disabled={store.graphInstance.forceEngine}
+                        />
+                    </Tooltip>
+                </HStack>
+            </VStack>
+        );
+    };
+
+    return (
+        <VStack
+            align="center"
+            direction="column"
+            paddingLeft="0"
+            paddingRight="0"
+            id="viewsettingscomponent"
+            width="100%"
+        >
+            <VStack spacing="2px" align="start" width="100%">
+                <CanvasTools />
+                <EdgeTools />
+                <NodeTools />
+                <FilterTools />
+            </VStack>
+            <VStack
+                width="100%"
+                backgroundColor="whiteAlpha.200"
+                padding="10px"
+                borderRadius="10px"
+                style={{ marginTop: '15px' }}
+            >
+                <Heading
+                    size="sm"
+                    style={{ marginBottom: '10px' }}
+                    width="100%"
+                >
+                    Layout
+                </Heading>
+                {renderLayoutOptions()}
+            </VStack>
+        </VStack>
+    );
+}
+
+export default observer(ViewTools);
